moved reducer for updating the user profile into fetchUserProfile.js. The page's state no longer has both a 'userProfile' and 'updateUserProfile' property. It still has the 'auth' property, which holds the exact same data as the 'userProfile' property, but doesn't get updated when using the website's functionality for updating a user's profile. More work needed to sort that out.
