+++ app/components/UserProfile.js
@@ -7,7 +7,6 @@ import UpdateProfile from "./UpdateProfile";
 const UserProfile = (props) => {
     const { id } = useParams(); 
     const { fetchUserProfile, userProfile, updateUserProfile } = props;
-    console.log(props);
 
     useEffect(() => {
         fetchUserProfile(id);
