+++ app/reducers/fetchUserProfile.js
@@ -1,6 +1,7 @@
 import axios from 'axios';
 
 const FETCH_USER_PROFILE = 'FETCH_USER_PROFILE';
+const UPDATE_USER_PROFILE = "UPDATE USER PROFILE";
 
 //action creators
 
@@ -24,6 +25,17 @@ export const fetchUserProfileThunk = (id) => {
     }
 }
 
+export function updateUserProfileThunk(id, update){
+    return async dispatch => {
+        try {
+            const user = await axios.put(`api/users/:${id}`, update);
+            dispatch({type: UPDATE_USER_PROFILE, update})
+        } catch (error) {
+            console.log("Invalid");
+        }
+    }
+}
+
 const initialState = {};
 
 //reducers
@@ -32,6 +44,8 @@ export default function (state = initialState, action) {
     switch (action.type) {
         case FETCH_USER_PROFILE:
             return action.userProfile;
+        case UPDATE_USER_PROFILE:
+            return {...state, ...action.update};
         default:
             return state;
     }
